+++ tests/components/openwater/dummy.py
@@ -62,13 +62,14 @@ class Dummy(OpenWaterComponent):
     }
     _constants_info = {
         'constant_c': {
-            'units': '1'
+            'units': '1',
+            'default_value': 3
         }
     }
     _states_info = {
         'state_a': {
             'units': '1',
-            'divisions': 1
+            'divisions': (4, 'constant_c')
         }
     }
     _outputs_info = {
@@ -84,6 +85,8 @@ class Dummy(OpenWaterComponent):
     def initialise(self,
                    # component states
                    state_a,
+                   # component constants
+                   constant_c,
                    **kwargs):
 
         state_a[-1][:] = 0
@@ -99,7 +102,7 @@ class Dummy(OpenWaterComponent):
             # component states
             state_a,
             # component constants
-            constant_c=3,
+            constant_c,
             **kwargs):
 
         state_a[0][:] = state_a[-1] + 1
@@ -107,14 +110,16 @@ class Dummy(OpenWaterComponent):
         return (
             # to exchanger
             {
-                'transfer_l': ancillary_b[11] * transfer_m + state_a[0],
+                'transfer_l':
+                    ancillary_b[11] * transfer_m + state_a[0][..., 0, 0],
                 'transfer_n': parameter_c * transfer_j,
                 'transfer_o': constant_c + transfer_j
             },
             # component outputs
             {
                 'output_x': parameter_c * transfer_j + constant_c,
-                'output_y': ancillary_b[11] * transfer_m - state_a[0],
+                'output_y':
+                    ancillary_b[11] * transfer_m - state_a[0][..., 0, 0]
             }
         )
 
@@ -130,7 +135,7 @@ class DummyFortran(Dummy):
     _states_info = {
         'state_a': {
             'units': '1',
-            'divisions': 1,
+            'divisions': (4, 'constant_c'),
             'order': 'F'
         }
     }
@@ -138,8 +143,10 @@ class DummyFortran(Dummy):
     def initialise(self,
                    # component states
                    state_a,
+                   # component constants
+                   constant_c,
                    **kwargs):
-        dummyfortran.initialise(state_a[-1])
+        dummyfortran.initialise(state_a[-1], constant_c=constant_c)
 
     def run(self,
             # from exchanger
@@ -152,13 +159,13 @@ class DummyFortran(Dummy):
             # component states
             state_a,
             # component constants
-            constant_c=3,
+            constant_c,
             **kwargs):
 
         transfer_l, transfer_n, transfer_o, output_x, output_y = (
             dummyfortran.run(
                 transfer_j, transfer_m, ancillary_b, parameter_c,
-                state_a[-1], state_a[0], constant_c
+                state_a[-1], state_a[0], constant_c=constant_c
             )
         )
 
@@ -188,8 +195,10 @@ class DummyC(Dummy):
     def initialise(self,
                    # component states
                    state_a,
+                   # component constants
+                   constant_c,
                    **kwargs):
-        dummyc.initialise(state_a[-1])
+        dummyc.initialise(constant_c, state_a[-1])
 
     def run(self,
             # from exchanger
@@ -202,7 +211,7 @@ class DummyC(Dummy):
             # component states
             state_a,
             # component constants
-            constant_c=3,
+            constant_c,
             **kwargs):
 
         transfer_l, transfer_n, transfer_o, output_x, output_y = (