+++ tests/components/openwater/dummyc/dummy.c
@@ -1,19 +1,28 @@
 void initialise_(int nz, int ny, int nx,
+                 // component constants,
+                 double constant_c,
                  // component states
                  double *state_a_m1)
 {
-  int i, j, k;
-  int ijk;
+  int i, j, k, l, m;
+  int nv, nw;
+  int ijklm;
+
+  // dimensions for state division
+  nw = 4;
+  nv = constant_c;
 
   for (i=0; i < nz; i++)
     for (j=0; j < ny; j++)
       for (k=0; k < nx; k++)
-      {
-        // vectorisation of 3d-array
-        ijk = k + nx * (j + ny * i);
-        // initialise states
-        state_a_m1[ijk] = 0.0;
-      }
+        for (l=0; l < nw; l++)
+          for (m=0; m < nv; m++)
+          {
+            // vectorisation of 5d-array
+            ijklm = m + nv * (l + nw * (k + nx * (j + ny * i)));
+            // initialise states
+            state_a_m1[ijklm] = 0.0;
+          }
 }
 
 void run_(int nz, int ny, int nx,
@@ -22,7 +31,7 @@ void run_(int nz, int ny, int nx,
           // component ancillary data
           double *ancillary_b,
           // component parameters
-          double parameter_c,
+          double *parameter_c,
           // component states
           double *state_a_m1, double *state_a_0,
           // component constants,
@@ -32,11 +41,16 @@ void run_(int nz, int ny, int nx,
           // component outputs
           double *output_x, double *output_y)
 {
-  int h, i, j, k;
-  int hijk, ijk;
+  int h, i, j, k, l, m;
+  int nv, nw;
+  int ijklm, hijk, ijk;
 
   // time dimension for monthly ancillary
   h = 11;
+
+  // dimensions for state division
+  nw = 4;
+  nv = constant_c;
 
   for (i=0; i < nz; i++)
     for (j=0; j < ny; j++)
@@ -47,16 +61,27 @@ void run_(int nz, int ny, int nx,
         // vectorisation of 3d-array
         ijk = k + nx * (j + ny * i);
         // update states
-        state_a_0[ijk] = state_a_m1[ijk] + 1;
+        for (l=0; l < nw; l++)
+          for (m=0; m < nv; m++)
+          {
+            // vectorisation of 5d-array
+            ijklm = m + nv * (l + nw * (k + nx * (j + ny * i)));
+            // initialise states
+            state_a_0[ijklm] = state_a_m1[ijklm] + 1;
+          }
+        // vectorisation of 5d-array
+        l = 0;
+        m = 0;
+        ijklm = m + nv * (l + nw * (k + nx * (j + ny * i)));
         // compute transfers to exchanger
         transfer_l[ijk] = (ancillary_b[ijk] * transfer_m[ijk])
-          + state_a_0[ijk];
-        transfer_n[ijk] = parameter_c * transfer_j[ijk];
-        transfer_o[ijk] = parameter_c + transfer_j[ijk];
+          + state_a_0[ijklm];
+        transfer_n[ijk] = parameter_c[ijk] * transfer_j[ijk];
+        transfer_o[ijk] = parameter_c[ijk] + transfer_j[ijk];
         // compute outputs
-        output_x[ijk] = (parameter_c * transfer_j[ijk]) + constant_c;
+        output_x[ijk] = (parameter_c[ijk] * transfer_j[ijk]) + constant_c;
         output_y[ijk] = (ancillary_b[ijk] * transfer_m[ijk])
-          - state_a_0[ijk];
+          - state_a_0[ijklm];
       }
 }