+++ tests/components/openwater/dummyfortran/dummy.f90
@@ -1,10 +1,12 @@
-subroutine initialise(z, y, x, state_a_m1)
+subroutine initialise(z, y, x, constant_c, state_a_m1)
     implicit none
 
     ! spaceshape
     integer, intent(in) :: z, y, x
+    ! component constants
+    integer, intent(in) :: constant_c
     ! component states
-    real(kind=8), intent(inout), dimension(z, y, x) :: state_a_m1
+    real(kind=8), intent(inout), dimension(z, y, x, 4, constant_c) :: state_a_m1
 
     state_a_m1 = 0
 
@@ -28,12 +30,12 @@ subroutine run(z, y, x, &
     ! component ancillary data
     real(kind=8), intent(in), dimension(12, z, y, x) :: ancillary_b
     ! component parameters
-    real(kind=8), intent(in) :: parameter_c
-    ! component states
-    real(kind=8), intent(in), dimension(z, y, x) :: state_a_m1
-    real(kind=8), intent(inout), dimension(z, y, x) :: state_a_0
+    real(kind=8), intent(in), dimension(z, y, x) :: parameter_c
     ! component constants
-    real(kind=8), intent(in) :: constant_c
+    integer, intent(in) :: constant_c
+    ! component states
+    real(kind=8), intent(in), dimension(z, y, x, 4, constant_c) :: state_a_m1
+    real(kind=8), intent(inout), dimension(z, y, x, 4, constant_c) :: state_a_0
     ! to exchanger
     real(kind=8), intent(out), dimension(z, y, x) :: &
         transfer_l, transfer_n, transfer_o
@@ -43,12 +45,12 @@ subroutine run(z, y, x, &
 
     state_a_0 = state_a_m1 + 1
 
-    transfer_l = (ancillary_b(12,:,:,:) * transfer_m) + state_a_0
+    transfer_l = (ancillary_b(12,:,:,:) * transfer_m) + state_a_0(:,:,:,1,1)
     transfer_n = parameter_c * transfer_j
     transfer_o = parameter_c + transfer_j
 
     output_x = (parameter_c * transfer_j) + constant_c
-    output_y = (ancillary_b(12,:,:,:) * transfer_m) - state_a_0
+    output_y = (ancillary_b(12,:,:,:) * transfer_m) - state_a_0(:,:,:,1,1)
 
 end subroutine run