+++ tests/components/surfacelayer/__init__.py
@@ -1,4 +1,1 @@
-
-
 from .dummy import Dummy, DummyFortran, DummyC
-from .sciencish import Sciencish