+++ changelog.rst
@@ -3,10 +3,24 @@ v0.1.0
 
 Released on 2021, ???.
 
+.. rubric:: API changes
+
+* add units requirement for component parameters and constants
+  (`#21 <https://github.com/hydro-jules/cm4twc/issues/21>`_)
+
 .. rubric:: Bug fixes
 
 * fix dump file update bug due to missing 'divisions' dimension
   (`#32 <https://github.com/hydro-jules/cm4twc/issues/32>`_)
+
+.. rubric:: Enhancements
+
+* add support for arrays for component parameters
+  (`#21 <https://github.com/hydro-jules/cm4twc/issues/21>`_)
+* add support for multiple divisions for component states
+  (`#39 <https://github.com/hydro-jules/cm4twc/pull/39>`_)
+* add support for customisable divisions for component states
+  (`#31 <https://github.com/hydro-jules/cm4twc/issues/31>`_)
 
 .. rubric:: Documentation